1code/oauth2-flask/server/requirements.txt
# alembic==1.7.6
apispec==5.1.1
# coverage==6.3.2
coverage
WTForms==2.3.1
Flask==2.0.1
flask_bootstrap==3.3.7.1
flask_marshmallow==0.14.0
flask_moment==1.0.2
flask_restplus==0.13.0
flask_restx==0.5.1
Flask_Cors==3.0.10
Flask_Admin==1.6.0
Flask_Migrate==3.1.0
Flask_Bootstrap==3.3.7.1
Flask_SQLAlchemy==2.5.1

requests==2.26.0
itsdangerous==2.0.1

marshmallow==3.14.1
six==1.14.0
# SQLAlchemy==1.4.31
# webargs==8.1.0
webargs
Werkzeug==2.0.3
# WTForms==3.0.1
# WTForms==2.3.1

MarkupSafe
Click
flask-pagedown


xmltodict
redis
cymysql==0.9.18
PyMySQL
mysql
mysql-connector-python


gunicorn

GitPython
